manager refactor: use Rc, and better manage state
We now manage state much better.

Instead of mutexes everywhere, we store Rc<RefCell<_>> everywhere.  This
should be better performance, especially since we're not using threads
everywhere anymore.

Additionally, the UnitManager is better organized now.  It is much
better about passing messages to various interfaces, and handles errors
more gracefully.

This is still largely able to run the complex Webserver Interface.  It
is not yet able to start scenarios, but it is well on its way to being
able to do so.

Signed-off-by: Sean Cross <sean@xobs.io>