logger: first attempt at adding logger support
This patch adds support for loggers.

Loggers aren't actually implemented yet, and it hasn't been tested, but
the unit files should load and they should be able to be activated.

They behave a lot like Interface files at the moment, which is a problem
that will be fixed by the next patch.

Signed-off-by: Sean Cross <sean@xobs.io>