Start scenario on terminal 'enter' press
This allows key presses in the terminal to be caught and then broadcast,
which allows an enter press to be used to start the scenario. This is
useful during development and prototyping. It could also be useful for
small systems where an external test trigger isn't available or
necessary.

This could be extended in the future to do more advanced things such as
printing the number of passed/failed boards since start up or to change
the verbosity of the printing.

Signed-off-by: Sam Lewis <sam.vr.lewis@gmail.com>